start extraction of user analytics
